[common] Renamed math_utils.hpp to mathutils.hpp
