+++ cudaaligner/src/myers_gpu.cu
@@ -14,7 +14,7 @@
 #include <vector>
 #include <numeric>
 #include <utils/signed_integer_utils.hpp>
-#include <utils/math_utils.hpp>
+#include <utils/mathutils.hpp>
 #include "device_storage.cuh"
 #include "batched_device_matrices.cuh"
 #include <cudautils/cudautils.hpp>
